Addressed some review comments and fixed a bug.
